requirements.txt
sqlalchemy
reflex
python-dotenv
email-validator
reflex==0.7.8